refactor: Change file structure of host
