+++ host/src/bin/docs.rs
